feat(seo): improve local SEO targeting San Francisco Bay Area smart contract market
- Update root layout metadata with Bay Area/SF location keywords and LocalBusiness JSON-LD schema
- Add geographic meta tags (geo.region, geo.placename, geo.position) for local search signals
- Enhance Organization schema with SF address, GeoCoordinates, and areaServed targeting
- Add FAQPage schema with common smart contract audit questions and answers
- Convert all client-component pages to server+client pattern to enable per-page metadata
- Add location-optimized metadata to services, about, contact, pricing, faq, and blog pages
- Create /services/san-francisco location landing page targeting Bay Area long-tail keywords
- Create two SEO-targeted blog articles:
  - "Top Smart Contract Vulnerabilities Found in Bay Area DeFi Projects"
  - "Web3 Security Checklist for Startups in the Bay Area"
- Update blog listing page to display published articles with proper linking
- Update sitemap.ts to include all new pages at appropriate priority levels

Targets long-tail keywords: "smart contract audit San Francisco", "EVM audit Bay Area",
"DeFi security audit San Francisco", "Web3 security services Bay Area"
